fix: web screen layout
